Initial commit
The files committed are code for a Netflix Clone from start to finish.
Git was not initializd at the start of the coding session and thus the changes were not tracked.
The code is written in React Js, deployed and hosted using Firebase hosting Service, by Google.
This is the link to where it is hosted : https://netflix-clone-20cbf.web.app/